allow multiple env flags, rename common file
